[cli] clean starter project (#85)
* [cli] init starter project remove .git

* [cli] init starter project remove .git tidy up

* [cli] init starter project remove .git

* [cli] init starter project remove .git tidy up

* [cli] Fix remove .git

* [cli] add test and when git clone if project directory exist should throw error

+++ packages/cli/src/controller/init-controller.ts
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import fs from 'fs';
+import rimraf from 'rimraf';
 import simpleGit, {SimpleGit} from 'simple-git';
 const git: SimpleGit = simpleGit();
 
@@ -9,15 +10,16 @@ const starterPath = 'https://github.com/OnFinality-io/subql-starter';
 
 export async function createProject(projectName: string): Promise<void> {
   const localPath = `${process.cwd()}/${projectName}`;
-  try {
-    await git.clone(starterPath, localPath);
-    const packageData = fs.readFileSync(`${localPath}/package.json`);
-    const currentPackage = JSON.parse(packageData.toString());
-    currentPackage.name = projectName;
-    const newPackage = JSON.stringify(currentPackage, null, 2);
-    fs.writeFileSync(`${localPath}/package.json`, newPackage, 'utf8');
-  } catch (e) {
-    /* handle all errors here */
-    console.log(e);
+  if (fs.existsSync(localPath)) {
+    throw new Error('Directory exists');
   }
+  await git.clone(starterPath, localPath);
+  const packageData = fs.readFileSync(`${localPath}/package.json`);
+  const currentPackage = JSON.parse(packageData.toString());
+  currentPackage.name = projectName;
+  const newPackage = JSON.stringify(currentPackage, null, 2);
+  fs.writeFileSync(`${localPath}/package.json`, newPackage, 'utf8');
+  await new Promise<void>((resolve, reject) =>
+    rimraf(`${localPath}/.git`, (error) => (error ? reject(error) : resolve()))
+  );
 }
